Implemented TagData sharing to reduce profile file size.
